Add files via upload
Due to climate change, food security and depletion of resources, there is a great need for green farming and innovation in South Africa. To address such challenges and bridge the gap between green technology providers and the agriculture sector, we have decided to develop a web platform called Agri-Energy connect.